+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -50,6 +50,7 @@ public let masterRuleList = RuleList(rules: [
     ForceTryRule.self,
     ForceUnwrappingRule.self,
     FunctionBodyLengthRule.self,
+    FunctionBodyWhitespaceLinesRule.self,
     FunctionParameterCountRule.self,
     GenericTypeNameRule.self,
     IdentifierNameRule.self,