+++ CHANGELOG.md
@@ -6,10 +6,16 @@
 
 #### Enhancements
 
+* Adds `function_body_whitespace_lines` opt-in rule
+  to prohibit whitespace lines in function bodies.  
+  [Mikhail Yakushin](https://github.com/driver733)
+  [#2065](https://github.com/realm/SwiftLint/issues/2065)
+
 * Add a new `excluded` config parameter to the `explicit_type_interface` rule
   to exempt certain types of variables from the rule.  
   [Rounak Jain](https://github.com/rounak)
   [#2028](https://github.com/realm/SwiftLint/issues/2028)
+
 
 #### Bug Fixes
 
@@ -23,7 +29,7 @@
 
 * None.
 
-#### Enhancements
+### Enhancements
 
 * Adds `discouraged_optional_boolean` opt-in rule to discourage
   the use of optional booleans.  
@@ -89,7 +95,10 @@
 
 #### Enhancements
 
-* None.
+* Adds `function_body_whitespace_comment` opt-in rule
+  to prohibit empty and comment lines in function bodies.  
+  [Mikhail Yakushin](https://github.com/driver733)
+  [#2065](https://github.com/realm/SwiftLint/issues/2065)
 
 #### Bug Fixes